+++ crates/azure_types/src/accounts.rs
@@ -7,12 +7,12 @@ pub struct Account {
     #[facet(rename = "cloudName")]
     pub cloud_name: String,
     #[facet(rename = "homeTenantId")]
-    pub home_tenant_id: AzureTenantId,
+    pub home_tenant_id: Option<AzureTenantId>,
     pub id: Uuid,
     #[facet(rename = "isDefault")]
     pub is_default: bool,
     #[facet(rename = "managedByTenants")]
-    pub managed_by_tenants: Vec<RawJson<'static>>,
+    pub managed_by_tenants: Option<Vec<RawJson<'static>>>,
     pub name: String,
     pub state: String,
     #[facet(rename = "tenantId")]
@@ -29,4 +29,17 @@ pub struct TenantIdHolder {
 #[derive(Debug, Eq, PartialEq, facet::Facet)]
 pub struct AccountUser {
     pub name: String,
+    #[facet(rename = "type")]
+    pub r#type: AccountUserType,
+}
+
+#[derive(Debug, Eq, PartialEq, facet::Facet)]
+#[repr(C)]
+pub enum AccountUserType {
+    #[facet(rename = "user")]
+    User,
+    #[facet(rename = "servicePrincipal")]
+    ServicePrincipal,
+    #[facet(other)]
+    Other(String),
 }